Se crearon algunos controllers

+++ src/main/java/co/edu/uniquindio/tienda/model/DetalleVenta.java
@@ -13,5 +13,5 @@ public class DetalleVenta {
 
     private Producto producto;
     private Integer cantidad;
-    private Integer subTotal;
+    private Double subTotal;
 }
